% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAM.R
\name{checkFull}
\alias{checkFull}
\title{Check whether the sample set is full}
\usage{
checkFull(split.info, num.train, num.test, num.valid)
}
\arguments{
\item{split.info}{List type, which contains the original data set, three sampling subsets, termination signal and other relevant sampling information.}

\item{num.train}{The number of training data points specified by the user.}

\item{num.test}{The number of test data points specified by the user.}

\item{num.valid}{The number of validation data points specified by the user.}
}
\value{
A list with sampling information.
}
\description{
Built-in function: This function includes four arguments,
where the first one contains the information of the original dataset as well as the three subsets,
and the remaining three augments are the maximum sample sizes for the training, test and validation subsets respectively.
}
