% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAM.R
\name{DP.reSample}
\alias{DP.reSample}
\title{Repeat sampling of DUPLEX}
\usage{
DP.reSample(split.info, choice)
}
\arguments{
\item{split.info}{A list containing relevant sampling information such as the original dataset and three sample subsets.}

\item{choice}{The variable must be one name of the three sample subsets contained in split.info, according to which the function assigns the current two data points to the specific sampling subset.}
}
\value{
Return the training, test and validation subsets. If the original data are required to be split into two subsets, the training and test subsets can be combined into a single calibration subset.
}
\description{
Built-in function: The cyclic sampling function of DUPLEX algorithm that takes the two data points farthest from the current sampling set and assigns them to the corresponding sampling subset.
}
