% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DSA_test_modData}
\alias{DSA_test_modData}
\title{Test data - middle}
\format{
A data frame with 1000 rows and 5 variables
#' \describe{
\item{Idex}{Data subscript that marks the position of each data point}
\item{I}{input vectors}
\item{I.1}{input vectors}
\item{I.2}{input vectors}
\item{O}{The output vector, usually the runoff}
...
}
}
\usage{
DSA_test_modData
}
\description{
A dataset containing the rainfall and runoff time series using for
testing data splitting algorithms
}
\keyword{datasets}
