% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DSAM_test_largeData}
\alias{DSAM_test_largeData}
\title{large test dataset}
\format{
A data frame with 3650 rows and 5 variables
\describe{
\item{Idex}{Data subscript that marks the position of each data point}
\item{I}{input vectors}
\item{I.1}{input vectors}
\item{I.2}{input vectors}
\item{O}{The output vector, usually the runoff}
...
}
}
\usage{
DSAM_test_largeData
}
\description{
A large dataset containing the rainfall and runoff time series using for
testing data splitting algorithms
}
\keyword{datasets}
