% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAM.R
\name{DUPLEX}
\alias{DUPLEX}
\title{'DSAM' - DUPLEX algorithm}
\usage{
DUPLEX(data, control)
}
\arguments{
\item{data}{The dataset should be matrix or Data.frame. The format should be as follows: Column one is a subscript vector used to mark each data point (each row is considered as a data point); Columns from 2 to N-1 are the input data, and Column N are the output data.}

\item{control}{User-defined parameter list, where each parameter definition refers to the \code{\link{par.default}} function.}
}
\value{
Return the training, test and validation subsets. If the original data are required to be split into two subsets, the training and test subsets can be combined into a single calibration subset.
}
\description{
The deterministic DUPLEX algorithm, with details given in Chen et al. (2022).
}
