% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAM.R
\name{getAUC}
\alias{getAUC}
\title{Get the AUC value between two datasets}
\usage{
getAUC(data1, data2)
}
\arguments{
\item{data1}{Dataset 1, the data type must be numeric, matrix or Data.frame.}

\item{data2}{Dataset 2, the data type must be numeric, matrix or Data.frame.}
}
\value{
Return the AUC value.
}
\description{
This function calls \code{[kohonen]{xgboost}} to train the classifier, followed by calculating the similarity between the two given datasets. The return value is a AUC index, ranging between 0 and 1, where the AUC is closer to 0.5, the more similar the two data sets is.
}
