% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAM.R
\name{getMax}
\alias{getMax}
\title{Get the maximum of the output column from the original data set}
\usage{
getMax(data)
}
\arguments{
\item{data}{The original data set, the data type must be numeric, matrix or Data.frame.}
}
\value{
Return the maximum value of the output column.
}
\description{
This function return the maximum of runoff(output columu) for users.
}
