% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAM.R
\name{getMean}
\alias{getMean}
\title{Get the mean and  standard deviation of the output column from the original data set}
\usage{
getMean(data)
}
\arguments{
\item{data}{The original data set, the data type must be numeric, matrix or Data.frame.}
}
\value{
Return a list with mean value and standard deviation.
}
\description{
This function return the mean and  standard deviation of runoff(output columu) for users.
}
