% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAM.R
\name{getMin}
\alias{getMin}
\title{Get the minimum of the output column from the original data set}
\usage{
getMin(data)
}
\arguments{
\item{data}{The original data set, the data type must be numeric, matrix or Data.frame.}
}
\value{
Return the minimum value of the output column.
}
\description{
This function return the minimum of runoff(output columu) for users.
}
