% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAM.R
\name{MDUPLEX}
\alias{MDUPLEX}
\title{'DSAM' - MDUPLEX algorithm}
\usage{
MDUPLEX(data, control)
}
\arguments{
\item{data}{The dataset should be matrix or Data.frame. The format should be as follows: Column one is a subscript vector used to mark each data point (each row is considered as a data point); Columns from 2 to N-1 are the input data, and Column N are the output data.}

\item{control}{User-defined parameter list, where each parameter definition refers to the par.default function.}
}
\value{
Return the training, test and validation subsets. If the original data are required to be split into two subsets, the training and test subsets can be combined into a single calibration subset.
}
\description{
This is a modified MDUPLEX algorithm, which is also deterministic, with details given in Zheng et al. (2022).
}
\references{
Chen, J., Zheng F., May R., Guo D., Gupta H., and Maier H. R.(2022), Improved data splitting methods for data-driven hydrological model development based on a large number of catchment samples, Journal of Hydrology, 613.

Zheng, F., Chen J., MaierH. R., and Gupta H.(2022), Achieving Robust and Transferable Performance for Conservation‐Based Models of Dynamical Physical Systems, Water Resources Research, 58(5).
}
