% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAM.R
\name{remainUnsample}
\alias{remainUnsample}
\title{Get the remain unsampled data after \code{\link{SSsample}}}
\usage{
remainUnsample(X, Y)
}
\arguments{
\item{X}{A vector that needs to be sampled.}

\item{Y}{A vector with data samples from X.}
}
\value{
A vector containing the remaining data that are not in Y.
}
\description{
Built-in function: This function is used in the semi-deterministic SS algorithm, and it contains two parameters X and Y, both of which are in an increased order. All data points in X vector that have not appeared in Y vector will be recorded and returned by this function.
}
