% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAM.R
\name{SBSS.P}
\alias{SBSS.P}
\title{'DSAM' - SBSS.P algorithm}
\usage{
SBSS.P(data, control)
}
\arguments{
\item{data}{The dataset should be matrix or Data.frame. The format should be as follows: Column one is a subscript vector used to mark each data point (each row is considered as a data point); Columns from 2 to N-1 are the input data, and Column N are the output data.}

\item{control}{User-defined parameter list, where each parameter definition refers to the par.default function.}
}
\value{
Return the training, test and validation subsets. If the original data are required to be split into two subsets, the training and test subsets can be combined into a single calibration subset.
}
\description{
SBSS.P algorithm is a stochastic algorithm. It obtains data subsets through uniform sampling in each neuron after clustering through SOM neural network, with details given in May et al. (2010).
}
\references{
May, R. J., Maier H. R., and Dandy G. C.(2010), Data splitting for artificial neural networks using SOM-based stratified sampling, Neural Netw, 23(2), 283-294.
}
