% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAM.R
\name{selectData}
\alias{selectData}
\title{Select specific split data}
\usage{
selectData(data, control)
}
\arguments{
\item{data}{The dataset should be matrix or Data.frame. The format should be as follows: Column one is a subscript vector used to mark each data point (each row is considered as a data point); Columns from 2 to N-1 are the input data, and Column N are the output data.}

\item{control}{User-defined parameter list, where each parameter definition refers to the par.default function.}
}
\value{
Returns a matrix for subsequent calculations.
}
\description{
Built-in function: This function decides whether to process the input dataset according to the parameter \code{include.inp}. If TRUE, this function removes Column 1 of the input dataset; otherwise, it returns the Column N of the data set.
}
