% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAM.R
\name{somCluster}
\alias{somCluster}
\title{Self-organized map clustering}
\usage{
somCluster(data)
}
\arguments{
\item{data}{The dataset in matrix or data.frame, containing only input and output vectors, but with no subscript vector.}
}
\value{
Return a data list of clustering neurons in the SOM network.
}
\description{
Built-in function: This function performs clustering for a given dataset by calling the \code{[kohonen]{som}} function from a “kohonen” package.
}
