% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAM.R
\name{SS}
\alias{SS}
\title{'DSAM' - SS algorithm}
\usage{
SS(data, control)
}
\arguments{
\item{data}{The type of data set to be divided should be matrix or Data.frame, and the data format is as follows: The first column is a subscript vector, which is used to mark each data point (each row is regarded as a data point); Columns 2 through N-1 are the input vectors, and columns N (the last) are the output vectors.}

\item{control}{User-defined parameter list, where each parameter definition refers to the \code{\link{par.default}} function.}
}
\value{
Return the training, test and validation subsets. If the original data are required to be split into two subsets, the training and test subsets can be combined into a single calibration subset.
}
\description{
The systematic stratified (SS) is a semi-deterministic method, with details given in Zheng et al. (2018).
}
\references{
Zheng, F., Maier, H.R., Wu, W., Dandy, G.C., Gupta, H.V. and Zhang, T. (2018) On Lack of Robustness in Hydrological Model Development Due to Absence of Guidelines for Selecting Calibration and Evaluation Data: Demonstration for Data‐Driven Models. Water Resources Research 54(2), 1013-1030.
}
