% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAM.R
\name{standardise}
\alias{standardise}
\title{Standardized data}
\usage{
standardise(data)
}
\arguments{
\item{data}{The dataset should be of type matrix or Data.frame and contain only the input and output vectors.}
}
\value{
Return a matrix with normalized data.
}
\description{
Built-in function: This function is used to standardize the data.
}
