% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAM.R
\name{TIMECON}
\alias{TIMECON}
\title{'DSAM' - Time-consecutive algorithm}
\usage{
TIMECON(data, control)
}
\arguments{
\item{data}{The dataset should be matrix or Data.frame. The format should be as follows: Column one is a subscript vector used to mark each data point (each row is considered as a data point); Columns from 2 to N-1 are the input data, and Column N are the output data.}

\item{control}{User-defined parameter list, where each parameter definition refers to the \code{\link{par.default}} function.}
}
\value{
Return the calibration and validation subsets.
}
\description{
This function selects a time-consecutive data from the original data set as the calibration (training and test) subset, and the remaining data is taken as the evaluation subset.
}
